Introduce LensCamera class
